api/requirements.txt
fastapi==0.115.0
httpx==0.25.2
pydantic==2.7.4
pydantic-settings>=2.2.1,<3.0.0
google-generativeai==0.8.0
python-dotenv==1.0.0
